# docs/swagger.yaml
definitions:
  check.Result:
    properties:
      id:
        type: integer
      resp_code:
        type: integer
      resp_duration:
        type: integer
      service_id:
        type: integer
      time_cheched:
        type: string
    required:
    - id
    - resp_code
    - resp_duration
    - service_id
    type: object
  check.ResultRequest:
    properties:
      id:
        type: integer
      resp_code:
        type: integer
      resp_duration:
        type: integer
      service_id:
        type: integer
    required:
    - id
    - resp_code
    - resp_duration
    - service_id
    type: object
  register.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  service.Service:
    properties:
      created:
        type: string
      id:
        type: integer
      interval:
        type: integer
      name:
        minLength: 3
        type: string
      url:
        type: string
    required:
    - id
    - interval
    - name
    - url
    type: object
  service.ServiceRequest:
    properties:
      id:
        type: integer
      interval:
        type: integer
      name:
        minLength: 3
        type: string
      url:
        type: string
    required:
    - id
    - interval
    - name
    - url
    type: object
info:
  contact: {}
paths:
  /api/result/item:
    post:
      consumes:
      - application/json
      parameters:
      - description: Модель результата провреки
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/check.ResultRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Result id already exists
          schema:
            type: string
        "401":
          description: invalid token
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Создание  результата провреки
      tags:
      - result
  /api/result/item/{id}:
    delete:
      parameters:
      - description: ID результата
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Item not found
          schema:
            type: string
        "401":
          description: Invalid token
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Удаление результата по ID
      tags:
      - result
    get:
      parameters:
      - description: ID результата
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/check.Result'
        "400":
          description: Item not found
          schema:
            type: string
      summary: Поиск результата по ID
      tags:
      - result
    put:
      consumes:
      - application/json
      parameters:
      - description: ID результата
        in: path
        name: id
        required: true
        type: integer
      - description: Модель результата
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/check.ResultRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Result item not found
          schema:
            type: string
        "401":
          description: Invalid token
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Изменение результата по ID
      tags:
      - result
  /api/result/items:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Item not found
          schema:
            type: string
      summary: Получить список всех результатов
      tags:
      - result
  /api/service/item:
    post:
      consumes:
      - application/json
      parameters:
      - description: Модель сервиса
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/service.ServiceRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Service id already exists
          schema:
            type: string
        "401":
          description: invalid token
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Создание сервиса
      tags:
      - service
  /api/service/item/{id}:
    delete:
      parameters:
      - description: ID сервиса
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Item not found
          schema:
            type: string
        "401":
          description: Invalid token
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Удаление сервиса по ID
      tags:
      - service
    get:
      parameters:
      - description: ID сервиса
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.Service'
        "400":
          description: Item not found
          schema:
            type: string
      summary: Поиск сервиса по ID
      tags:
      - service
    put:
      consumes:
      - application/json
      parameters:
      - description: ID сервиса
        in: path
        name: id
        required: true
        type: integer
      - description: Модель сервиса
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/service.ServiceRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Item not found
          schema:
            type: string
        "401":
          description: Invalid token
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Изменение сервиса по ID
      tags:
      - service
  /api/service/items:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Item not found
          schema:
            type: string
      summary: Получить список всех сервисов
      tags:
      - service
  /api/sign_in:
    post:
      consumes:
      - application/json
      parameters:
      - description: Модель аутентификации
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/register.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Invalid username or password
          schema:
            type: string
      summary: Аутентификация
      tags:
      - auth
swagger: "2.0"